+++ CHANGELOG.md
@@ -1,2 +1,11 @@
+## v0.1.0-rc.2 (2023-03-17)
+
+##### Client.lua
+* Document Client package properties
+
+##### Network.lua
+* Fix errors when communicator config wasn't supplied with `events` or `functions`
+* Fix documentation error
+
 ## v0.1.0-rc.1 (2023-03-17)
 * Initial release